requirements.txt
Flask==1.1.2
h5py==2.10.0
numpy==1.19.2
tensorflow-cpu==2.4.1
urllib3==1.25.9
Werkzeug==1.0.1
opencv-python-headless==4.2.0.32
wheel==0.35
gunicorn==20.0.4
Jinja2==2.11.2
wikipedia==1.4.0
